feat(ios): shared LocalActionRouter + foreground marker dispatch (Refs #122)
Extracts the background AppIntent's pattern matcher into a shared
`LocalActionRouter` enum (resolve → marker | answer | noMatch), and wires
the foreground orchestrator to the same fast path. Native CALL/SMS/OPEN
markers go through a new `MarkerDispatcher` that opens iOS-native URL
schemes (tel:, sms:, scheme://); spoken answers go straight to TTS without
spinning up the agent loop.

Result: in-app voice "call mom" / "send a message to Fede" / "open Spotify"
launches the same Apple-compliant native action surface that the Action
Button Shortcut path already uses, without any harness round-trip.
