+++ 02_GIGI_APP/GIGI/GigiBackgroundTalkIntent.swift
@@ -1,16 +1,12 @@
 import AppIntents
-import Contacts
 import Foundation
-import UserNotifications
-#if canImport(UIKit)
-import UIKit
-#endif
 
 // MARK: - GigiBackgroundTalkIntent
 //
 // Background-running AppIntent that receives a transcribed user phrase from
-// an iOS Shortcut, sends it to the GIGI harness for processing, and returns
-// the answer as a spoken dialog the Shortcut can route through Speak Text.
+// an iOS Shortcut, runs it through the shared `LocalActionRouter`, falls
+// back to the Mac harness for everything the router doesn't claim, and
+// returns the answer (or marker) as a string the Shortcut speaks or routes.
 // The app is never brought to the foreground (`openAppWhenRun: false`) — iOS
 // wakes the process briefly to run `perform()` and lets it return.
 //
@@ -22,421 +18,13 @@ import UIKit
 //      foregrounding.
 //   3. Shortcut → Run "Process speech with GIGI" with the dictated text.
 //      iOS calls into this intent in the background.
-//   4. We forward the text through `GigiHarnessClient.agentRun(text:)`. The
-//      harness performs whatever system-action routing or Claude reasoning
-//      it already does for the foreground app — same endpoint, same auth,
-//      same pairing. Nothing duplicated client-side.
-//   5. Shortcut → Speak Text on the dialog returned here.
+//   4. The router resolves system queries on-device. Anything else falls
+//      through to `GigiHarnessClient.agentRun`.
+//   5. Shortcut → Speak Text on the dialog returned here, or routes the
+//      `CALL:` / `SMS:` / `OPEN:` marker through its native action branches.
 //
-// Failure modes are spoken back through the same dialog channel so the user
-// always hears something even if the harness is offline or unpaired.
-
-// MARK: - LocalAnswer
-//
-// Pattern-based router that handles a small set of system queries entirely
-// on-device, without going through the harness. The point of this layer is
-// twofold:
-//   1. The banner stays useful when the Mac harness is unreachable (Mac
-//      off, tunnel down, not paired) for any request that doesn't need
-//      Claude or an external integration.
-//   2. We don't burn a Claude turn on questions whose answer is in the
-//      device clock or the current locale.
-//
-// Anything not matched here falls through to `agentRun`, which is where
-// reasoning, memory, and cross-platform actions (order pizza, book Uber,
-// search Amazon — the things that genuinely need the harness) are handled.
-//
-// Phrases are matched on lowercased / whitespace-trimmed text. We accept
-// English and Italian forms because the demo speaker mixes them. This is
-// deliberately a thin keyword router rather than the full GigiNLUEngine,
-// because the engine has UI dependencies that don't make sense in a
-// background AppIntent context.
-
-private enum LocalAnswer {
-    static func tryAnswer(for raw: String) async -> String? {
-        let lower = raw.lowercased().trimmingCharacters(in: .whitespacesAndNewlines)
-        guard !lower.isEmpty else { return nil }
-
-        if matchesAny(lower, prefixes: ["hello", "hi gigi", "hey gigi", "ciao", "ciao gigi"]) {
-            return "Hi! What can I help with?"
-        }
-
-        // Battery level — UIDevice exposes this from a background AppIntent
-        // once we flip the monitoring flag. Drains nothing since we read
-        // once and let iOS turn the monitor back off when the process is
-        // suspended.
-        if containsAny(lower, phrases: ["battery", "batteria", "carica", "how much battery"]) {
-            #if canImport(UIKit)
-            UIDevice.current.isBatteryMonitoringEnabled = true
-            let level = UIDevice.current.batteryLevel
-            if level >= 0 {
-                let percent = Int((level * 100).rounded())
-                return "Battery is at \(percent)%."
-            }
-            #endif
-            return "I can't read the battery level right now."
-        }
-
-        // Timer / reminder. Scheduling a UNNotificationRequest works fully
-        // in the background AppIntent context — no foreground required, no
-        // harness, no network. We parse a couple of common phrasings and
-        // confirm the result via dialog.
-        if let minutes = parseTimerMinutes(from: lower) {
-            scheduleLocalReminder(after: TimeInterval(minutes * 60),
-                                  body: "Timer expired.")
-            return "Timer set for \(minutes) minute\(minutes == 1 ? "" : "s")."
-        }
-
-        if let payload = parseReminder(from: raw) {
-            scheduleLocalReminder(after: 60, body: payload)
-            return "I'll remind you in a minute: \(payload)"
-        }
-
-        // Native actions that iOS refuses to launch reliably from a background
-        // AppIntent are returned as command markers. The user's Shortcut owns
-        // the privileged foreground action. For calls, the demo-safe path is:
-        //
-        //   "call mom" → resolve Mom in Contacts here → CALL:+15551234
-        //             → Shortcut Call action → iOS native call confirmation
-        //
-        // That keeps GIGI closed even if the user started from Instagram, while
-        // still using Apple's compliant call-confirmation surface. `CALL:` is
-        // intentionally left as a legacy/unresolved fallback for manually-built
-        // Shortcuts that prefer Get Contact + Call.
-        if let name = parseCallRequest(lower) {
-            return await callMarker(for: name)
-        }
-
-        if let message = parseMessageRequest(raw: raw, lower: lower) {
-            return await messageMarker(for: message)
-        }
-
-        if let target = parseOpenAppRequest(lower) {
-            return "OPEN:\(target.scheme)"
-        }
-
-        if containsAny(lower, phrases: ["what time", "what's the time", "che ore sono", "che ora è", "ora è"]) {
-            let formatter = DateFormatter()
-            formatter.timeStyle = .short
-            formatter.dateStyle = .none
-            return "It's \(formatter.string(from: Date()))."
-        }
-
-        if containsAny(lower, phrases: ["what day", "what's the date", "today's date", "che giorno è", "che data è"]) {
-            let formatter = DateFormatter()
-            formatter.dateStyle = .full
-            formatter.timeStyle = .none
-            return "Today is \(formatter.string(from: Date()))."
-        }
-
-        if containsAny(lower, phrases: ["thank you", "thanks", "grazie"]) {
-            return "You're welcome."
-        }
-
-        // Capability questions. Without this branch every "what can you do"
-        // hits the harness, which is wasteful and breaks completely if the
-        // harness isn't paired (the only feedback the user gets is the
-        // not-configured error). Hard-coding the capability list locally
-        // turns it into a fast, always-on answer.
-        if containsAny(lower, phrases: [
-            "what can you do",
-            "what do you do",
-            "what are you",
-            "who are you",
-            "help",
-            "capabilities",
-            "cosa sai fare",
-            "cosa puoi fare",
-            "chi sei",
-            "aiuto",
-            "cosa fai"
-        ]) {
-            return "I can answer questions, send messages, plan your day, and run actions on your devices. Try saying: what time is it, plan my day, send a message to Marco, or order a pizza."
-        }
-
-        // Self-test phrase the demo can use to confirm the banner is wired
-        // up before turning the harness on. Useful during stage prep.
-        if containsAny(lower, phrases: ["are you there", "ping", "test"]) {
-            return "I'm here. The banner works."
-        }
-
-        return nil
-    }
-
-    private static func matchesAny(_ text: String, prefixes: [String]) -> Bool {
-        prefixes.contains { text == $0 || text.hasPrefix($0 + " ") || text.hasPrefix($0 + ",") }
-    }
-
-    private static func containsAny(_ text: String, phrases: [String]) -> Bool {
-        phrases.contains { text.contains($0) }
-    }
-
-    // MARK: - Message marker
-
-    private enum MessagePlatform {
-        case sms
-        case whatsapp
-    }
-
-    /// Recognizes simple message requests for the Shortcut's Send Message branch.
-    /// Format returned to Shortcuts: `SMS:<phone>|<body>` for iMessage/SMS, or
-    /// `OPEN:whatsapp://send?...` for WhatsApp.
-    private static func parseMessageRequest(raw: String, lower: String) -> (contact: String, body: String, platform: MessagePlatform)? {
-        let triggers: [(prefix: String, platform: MessagePlatform)] = [
-            ("send a message to ", .sms),
-            ("send a text to ", .sms),
-            ("text ", .sms),
-            ("message ", .sms),
-            ("send an sms to ", .sms),
-            ("sms ", .sms),
-            ("whatsapp ", .whatsapp),
-            ("send a whatsapp to ", .whatsapp),
-            ("manda un messaggio a ", .sms),
-            ("scrivi a ", .sms),
-            ("messaggia ", .sms),
-            ("manda whatsapp a ", .whatsapp),
-            ("manda un whatsapp a ", .whatsapp)
-        ]
-        guard let match = triggers.first(where: { lower.hasPrefix($0.prefix) }) else { return nil }
-        let trigger = match.prefix
-        let restRaw = String(raw.dropFirst(trigger.count)).trimmingCharacters(in: .whitespacesAndNewlines)
-        let restLower = String(lower.dropFirst(trigger.count)).trimmingCharacters(in: .whitespacesAndNewlines)
-        guard !restRaw.isEmpty else { return nil }
-
-        let bodySeparators = [" saying ", " that ", " with message ", " with the message ", " dicendo ", " che "]
-        if let separator = bodySeparators.first(where: { restLower.contains($0) }),
-           let range = restLower.range(of: separator) {
-            let contactRaw = String(restRaw.prefix(restLower.distance(from: restLower.startIndex, to: range.lowerBound)))
-                .trimmingCharacters(in: .whitespacesAndNewlines)
-            let bodyStart = restLower.distance(from: restLower.startIndex, to: range.upperBound)
-            let bodyRaw = String(restRaw.dropFirst(bodyStart)).trimmingCharacters(in: .whitespacesAndNewlines)
-            if !contactRaw.isEmpty { return (cleanContactName(contactRaw), bodyRaw, match.platform) }
-        }
-
-        return (cleanContactName(restRaw), "", match.platform)
-    }
-
-    private static func messageMarker(for message: (contact: String, body: String, platform: MessagePlatform)) async -> String {
-        let recipient: String
-        if await ensureContactsAccessForBackgroundAction(),
-           let resolved = await GigiContactsEngine.shared.resolve(message.contact),
-           let phone = shortcutDialablePhoneNumber(from: resolved.phone) {
-            recipient = phone
-        } else {
-            recipient = message.contact
-        }
-
-        switch message.platform {
-        case .sms:
-            return "SMS:\(recipient)|\(message.body)"
-        case .whatsapp:
-            guard let phone = whatsappPhoneNumber(from: recipient) else {
-                return "SMS:\(recipient)|\(message.body)"
-            }
-            let encoded = urlQueryValue(message.body)
-            return "OPEN:whatsapp://send?phone=\(phone)&text=\(encoded)"
-        }
-    }
-
-    // MARK: - Open external app
-
-    /// Maps "open Spotify", "apri WhatsApp", etc. to a known URL scheme.
-    /// The list mirrors LSApplicationQueriesSchemes in Info.plist — adding
-    /// schemes here without listing them there makes canOpenURL return
-    /// false silently.
-    private static func parseOpenAppRequest(_ lower: String) -> (scheme: String, label: String)? {
-        let triggers = ["open ", "launch ", "start ", "apri ", "lancia ", "avvia "]
-        guard let trigger = triggers.first(where: { lower.hasPrefix($0) }) else { return nil }
-        let target = String(lower.dropFirst(trigger.count))
-            .trimmingCharacters(in: .whitespacesAndNewlines)
-            .trimmingCharacters(in: CharacterSet(charactersIn: "."))
-        if target.isEmpty { return nil }
-
-        let schemes: [(label: String, keywords: [String], scheme: String)] = [
-            ("Spotify",      ["spotify"],                     "spotify://"),
-            ("WhatsApp",     ["whatsapp", "whats app"],       "whatsapp://"),
-            ("Instagram",    ["instagram", "ig"],             "instagram://"),
-            ("Telegram",     ["telegram", "tg"],              "tg://"),
-            ("YouTube",      ["youtube", "yt"],               "youtube://"),
-            ("TikTok",       ["tiktok", "tik tok"],           "tiktok://"),
-            ("Maps",         ["maps", "apple maps", "mappe"], "maps://"),
-            ("Google Maps",  ["google maps"],                 "comgooglemaps://"),
-            ("Waze",         ["waze"],                        "waze://"),
-            ("Uber",         ["uber"],                        "uber://"),
-            ("Lyft",         ["lyft"],                        "lyft://"),
-            ("Uber Eats",    ["uber eats", "ubereats"],       "ubereats://"),
-            ("DoorDash",     ["doordash", "door dash"],       "doordash://")
-        ]
-
-        if let match = schemes.first(where: { entry in
-            entry.keywords.contains(where: { target.contains($0) })
-        }) {
-            return (match.scheme, match.label)
-        }
-        return nil
-    }
-
-    // MARK: - Place a call
-
-    /// Recognizes natural "call X" / "chiama X" patterns and returns the name.
-    /// Includes common dictation mistakes seen in live tests: "col mom" /
-    /// "col mam" for "call mom".
-    private static func parseCallRequest(_ lower: String) -> String? {
-        let starts = [
-            "call ", "phone ", "dial ", "ring ",
-            "col ", "calla ",
-            "chiama ", "chiamo ", "chiami ", "telefona a "
-        ]
-        let embedded = [
-            "can you call ", "could you call ", "please call ",
-            "puoi chiamare ", "puoi chiamarmi ", "mi chiami ",
-            "per favore chiama "
-        ]
-
-        let name: String
-        if let trigger = starts.first(where: { lower.hasPrefix($0) }) {
-            name = String(lower.dropFirst(trigger.count))
-        } else if let trigger = embedded.first(where: { lower.contains($0) }),
-                  let range = lower.range(of: trigger) {
-            name = String(lower[range.upperBound...])
-        } else {
-            return nil
-        }
-
-        let cleaned = name
-            .trimmingCharacters(in: .whitespacesAndNewlines)
-            .trimmingCharacters(in: CharacterSet(charactersIn: "."))
-            .removingCallPolitenessSuffixes()
-        return cleaned.isEmpty ? nil : cleanContactName(cleaned)
-    }
-
-    private static func callMarker(for contactName: String) async -> String {
-        guard await ensureContactsAccessForBackgroundAction() else {
-            return "I need Contacts permission to call \(contactName). Open GIGI once, allow Contacts, then try again from the Action Button."
-        }
-
-        guard let resolved = await GigiContactsEngine.shared.resolve(contactName) else {
-            // Legacy fallback: if the user-built Shortcut has a Get Contact +
-            // Call branch it can still try iOS' own contact picker/resolver.
-            return "CALL:\(contactName)"
-        }
-
-        guard let phoneNumber = shortcutDialablePhoneNumber(from: resolved.phone) else {
-            return "I found \(resolved.name), but the phone number doesn't look dialable."
-        }
-
-        // Canonical demo marker. The Shortcut's CALL branch strips `CALL:` and
-        // passes this value into the native Shortcuts "Call" action. That is
-        // more reliable than Open URL + tel: and still keeps GIGI closed while
-        // iOS owns the compliant call confirmation surface.
-        return "CALL:\(phoneNumber)"
-    }
-
-    private static func ensureContactsAccessForBackgroundAction() async -> Bool {
-        switch CNContactStore.authorizationStatus(for: .contacts) {
-        case .authorized, .limited:
-            return true
-        case .denied, .restricted:
-            return false
-        case .notDetermined:
-            return await withCheckedContinuation { continuation in
-                CNContactStore().requestAccess(for: .contacts) { granted, _ in
-                    continuation.resume(returning: granted)
-                }
-            }
-        @unknown default:
-            return false
-        }
-    }
-
-    private static func shortcutDialablePhoneNumber(from phone: String) -> String? {
-        // Keep only real phone-number characters for the native Shortcuts Call
-        // action. We deliberately drop service-code characters for demo safety.
-        let cleaned = phone.filter { "0123456789+".contains($0) }
-        guard cleaned.filter(\.isNumber).count >= 3 else { return nil }
-        return cleaned
-    }
-
-    private static func whatsappPhoneNumber(from phone: String) -> String? {
-        let digits = phone.filter(\.isNumber)
-        return digits.count >= 6 ? digits : nil
-    }
-
-    private static func urlQueryValue(_ value: String) -> String {
-        var allowed = CharacterSet.urlQueryAllowed
-        allowed.remove(charactersIn: "+&=#?")
-        return value.addingPercentEncoding(withAllowedCharacters: allowed) ?? value
-    }
-
-    private static func cleanContactName(_ name: String) -> String {
-        name.trimmingCharacters(in: .whitespacesAndNewlines)
-            .trimmingCharacters(in: CharacterSet(charactersIn: ".,!?"))
-            .split(separator: " ")
-            .map { $0.prefix(1).uppercased() + $0.dropFirst() }
-            .joined(separator: " ")
-    }
-
-    /// Pulls a minute count out of common timer phrasings.
-    /// Examples it handles: "set a timer for 5 minutes", "timer 10",
-    /// "metti un timer di 3 minuti", "timer di 7".
-    private static func parseTimerMinutes(from lower: String) -> Int? {
-        let triggers = ["timer", "set a timer", "start a timer", "metti un timer", "imposta un timer"]
-        guard triggers.contains(where: lower.contains) else { return nil }
-        // Find the first integer in the string.
-        let scanner = Scanner(string: lower)
-        scanner.charactersToBeSkipped = .alphanumerics.subtracting(.decimalDigits).union(.whitespaces).union(.punctuationCharacters)
-        var value: Int = 0
-        if scanner.scanInt(&value), value > 0, value <= 600 {
-            return value
-        }
-        return nil
-    }
-
-    /// Strips a leading "remind me to" / "ricordami di" prefix and returns
-    /// the remaining body, which becomes the notification text.
-    private static func parseReminder(from raw: String) -> String? {
-        let lower = raw.lowercased()
-        for prefix in ["remind me to ", "remind me ", "ricordami di ", "ricordami "] {
-            if lower.hasPrefix(prefix) {
-                let body = String(raw.dropFirst(prefix.count)).trimmingCharacters(in: .whitespacesAndNewlines)
-                if !body.isEmpty { return body }
-            }
-        }
-        return nil
-    }
-
-    /// Schedules a local notification N seconds from now. UNUserNotificationCenter
-    /// is reachable from a background AppIntent; the system fires the alert even
-    /// after the GIGI process exits. If the user has never granted notification
-    /// permission this returns silently — the dialog above will still play, the
-    /// notification just won't appear, which is acceptable degradation.
-    private static func scheduleLocalReminder(after seconds: TimeInterval, body: String) {
-        let content = UNMutableNotificationContent()
-        content.title = "GIGI"
-        content.body = body
-        content.sound = .default
-        let trigger = UNTimeIntervalNotificationTrigger(timeInterval: max(1, seconds), repeats: false)
-        let request = UNNotificationRequest(identifier: UUID().uuidString, content: content, trigger: trigger)
-        UNUserNotificationCenter.current().add(request, withCompletionHandler: nil)
-    }
-}
-
-private extension String {
-    func removingCallPolitenessSuffixes() -> String {
-        var result = self.trimmingCharacters(in: .whitespacesAndNewlines)
-        let suffixes = [
-            " for me", " please", " now",
-            " per favore", " grazie", " adesso"
-        ]
-        for suffix in suffixes {
-            if result.hasSuffix(suffix) {
-                result = String(result.dropLast(suffix.count))
-                    .trimmingCharacters(in: .whitespacesAndNewlines)
-            }
-        }
-        return result
-    }
-}
+// Failure modes from the harness are mapped to user-friendly speech so the
+// user always hears something even when the Mac is unreachable.
 
 @available(iOS 16.0, *)
 struct GigiBackgroundTalkIntent: AppIntent {
@@ -470,7 +58,7 @@ struct GigiBackgroundTalkIntent: AppIntent {
         // unreachable (Mac off, tunnel down, not paired). Only requests
         // that need cross-platform actions (order, book, browse) or
         // language-model reasoning fall through to the harness path.
-        if let local = await LocalAnswer.tryAnswer(for: trimmed) {
+        if let local = await LocalActionRouter.tryAnswer(for: trimmed) {
             return .result(value: local)
         }
 
@@ -490,13 +78,6 @@ struct GigiBackgroundTalkIntent: AppIntent {
             let message: String
             switch err {
             case .notConfigured:
-                // The user invoked GIGI for something that needs the Mac
-                // harness (Claude reasoning, memory, cross-platform tools)
-                // but pairing hasn't been completed on this device. We
-                // can't pair from a background AppIntent, so we tell the
-                // user concretely what to do and what still works without
-                // the harness — otherwise the message reads as a generic
-                // "open the app" that gives no signal about why.
                 message = "I'm running without the Mac brain. Ask me about the time, the date, or say hello — that works on the phone alone. For everything else, open the GIGI app and pair it with your Mac."
             case .transport:
                 message = "I couldn't reach GIGI. Check the connection and try again."
